ci: consolidate duplicate Deno quality workflows (#231)
Merge `deno-quality.yml` into `quality.yml` so PRs run lint, fmt,
type-check, and tests once instead of twice. The merged workflow
keeps coverage + Codecov upload (issue #51) and SHA-pinned actions,
while still running the example programs.

- Delete `.github/workflows/deno-quality.yml` and its test file.
- Update `quality_test.ts` with assertions for coverage, Codecov,
  and SHA-pinning that previously lived in `deno_quality_test.ts`.
- Drop the deleted workflow from `workflow_permissions_test.ts`.

Closes #231.
