--- .github/workflows/deno-quality.yml
@@ -1,61 +1,0 @@
-# Deno Quality workflow for NEAT-AI-Examples
-#
-# Runs Deno lint, format check, type check, and unit tests with code
-# coverage on every pull request, then uploads the lcov report to
-# Codecov. Issue #51.
-
-name: Deno Quality
-
-on:
-  pull_request:
-    branches: ["*"]
-
-permissions:
-  contents: read
-
-jobs:
-  quality:
-    name: Lint, format, type-check, test
-    runs-on: ubuntu-latest
-    steps:
-      - name: Check out repository
-        uses: actions/checkout@34e114876b0b11c390a56381ad16ebd13914f8d5 # v4.3.1
-
-      - name: Install Deno
-        uses: denoland/setup-deno@667a34cdef165d8d2b2e98dde39547c9daac7282 # v2.0.4
-        with:
-          deno-version: v2.x
-
-      - name: Run deno lint
-        run: deno lint
-
-      - name: Run deno fmt --check
-        run: deno fmt --check
-
-      - name: Run deno check (type checking)
-        run: deno check **/*.ts
-
-      # Run tests with coverage. The example runner scripts and the
-      # native FFI library are not available in this minimal workflow,
-      # so we only execute the unit tests here. The full quality.yml
-      # workflow continues to run the example programs.
-      #
-      # --allow-net is required so the @stsoftware/neat-ai WASM
-      # activation module can fetch its remote payload from jsr.io;
-      # --allow-ffi is included so the optional native discovery
-      # library can load when present. Issue #73.
-      - name: Run unit tests with coverage
-        run: deno test --no-check --allow-read --allow-write --allow-env --allow-net --allow-ffi --coverage=cov_profile
-
-      - name: Generate lcov coverage report
-        run: deno coverage cov_profile --lcov --output=coverage.lcov
-
-      # The codecov-action is a no-op when no coverage file is found,
-      # and we set fail_ci_if_error: false so a transient Codecov
-      # outage does not block the PR.
-      - name: Upload coverage to Codecov
-        uses: codecov/codecov-action@0f8570b1a125f4937846a11fcfa3bcd548bd8c97 # v4.6.0
-        with:
-          files: coverage.lcov
-          token: ${{ secrets.CODECOV_TOKEN }}
-          fail_ci_if_error: false
